+++ src/settings/MappingProfiles/MappingProfilesContainer/MappingProfilesContainer.js
@@ -25,6 +25,8 @@ import {
   FIND_ALL_CQL,
 } from '../../../utils';
 import { NewMappingProfileFormRoute } from '../NewMappingProfileFormRoute';
+import { generateTransformationFieldsValues } from '../MappingProfilesForm/TransformationsField';
+import { mappingProfileTransformations } from '../MappingProfilesForm/TransformationsField/transformations';
 
 const customProperties = getMappingProfilesColumnProperties({
   columnWidths: { format: '70px' },
@@ -47,6 +49,12 @@ const sortMap = {
   updatedBy: 'userInfo.firstName userInfo.lastName',
 };
 
+const initialValues = {
+  recordTypes: [],
+  outputFormat: 'MARC',
+  transformations: generateTransformationFieldsValues(mappingProfileTransformations),
+};
+
 const MappingProfilesContainer = ({
   history,
   match,
@@ -67,6 +75,7 @@ const MappingProfilesContainer = ({
         render={props => (
           <NewMappingProfileFormRoute
             {...props}
+            initialValues={initialValues}
             onCancel={() => history.push(`${match.path}${location.search}`)}
             onSubmit={mutator.mappingProfiles.POST}
           />